Make operators use full-prefix notation
Previously, there were two kinds of operators: 'special' and 'multi'.

A 'special' operator would be prefix and look like '^true', for example.

A 'multi' operator would also be prefix, but use brackets; '(+ 1 2)',
for example.

Now all operators are prefix and called like functions (syntax), so
'^(true)', and '+(1, 2)'.

Further changes hope to eliminate operators entirely and just used
symbolic functions. It may be possible to have sugar-syntax on top of
these to allow calling in infix or prefix (no brackets) notation.